Add a servo-like device
This is a really crude version of the servo object on the
microcontroller. It only stores the value that we think we're on with
a resolution of eight bits. (We use 0-999 on the controller.)

It's limited to eight bits to emulate DMX.

The update rate is set to 40Hz, which also emulates DMX. This is set via
a crude use of usleep(), so it's not the hard real time we have on the
controller, but it's fine. Any Linux PC is faster than the CPU on the Pi
Pico. 😅